Fix: Libs loaded before ensure_depenencies can be run
This was a bug created by new refactor work that returns this to its previous behavior.
