+++ Launcher/LauncherScript/_lib/__init__.py
@@ -2,7 +2,7 @@
 
 from .parse_ansi import AnsiParser
 from .ordered_logger import OrderedLogger
-from .utils import RingMovingAverage, is_shift_held, ensure_dependencies
+from .dependencies import ensure_dependencies
 
 __all__ = [
     'AnsiParser',
@@ -11,3 +11,13 @@ __all__ = [
     'is_shift_held',
     'ensure_dependencies',
 ]
+
+
+def __getattr__(name: str):
+    if name == "RingMovingAverage":
+        from .utils import RingMovingAverage
+        return RingMovingAverage
+    if name == "is_shift_held":
+        from .utils import is_shift_held
+        return is_shift_held
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
